src/old/requirements_taxonomy.txt
pandas
plotly
pyyaml
dash
pymongo